openflow: add aggregate stats fetching
Implement a minimalistic openflow 1.0 client library that can connect to
a bridge management socket, send a openflow HELLO message and finally
send a NX_AGGREGATE query to get the bridge aggregated statistics.

This was reverse engineered with a combination of

	strace -e trace=sendto,recvfrom -s 65536 -x \
		ovs-ofctl dump-aggregate $bridge

And by following the code path in ovs-vswitchd with gdb to determine the
actual data structures that were used in the ovs code base.

Before writing this on my own, I tried using a few "openflow" libraries
but they were very complicated and not working most of the time. Most
importantly, none of them supported NX_AGGREGATE which was the only
feature I required.

Signed-off-by: Robin Jarry <rjarry@redhat.com>