appctl: add unixctl json rpc wrapper
The unixctl socket uses the JSON-RPC protocol and returns plain text.

Some commands support returning JSON-in-JSON instead of text-in-JSON but
none of the commands that we need at the moment have that support. We'll
do with plain text for now.

The go library implementation of the JSON-RPC codec does not handle
arguments according to what ovs expects. Implements our own version of
it which is a copy of the standard library with some minor changes
around how command parameters are encoded.

Signed-off-by: Robin Jarry <rjarry@redhat.com>