collector: add base framework
Add a Collector interface that will need to be implemented by all
collectors. Add a Metric type which is used to define abstract
prometheus metrics with their name, description, labels and type.

The collectors need a name to be enabled/disabled in the config. Every
metric needs a MetricSet to determine which metrics of the enabled
collectors will actually be exported.

Add a utility function to print metrics and a helper to use in the
collectors Describe() method.

Signed-off-by: Robin Jarry <rjarry@redhat.com>