main: add program entry point
Implement the main entry point of the exporter.

On startup, it will first check the command line arguments. If -l is
present, all supported metrics of all collectors (regardless of any
configuration) are printed and the program exits.

If no arguments are specified, all enabled collectors are added to
a prometheus registry which is then passed to a promhttp handler
provided by the prometheus client library.

Implement a thin wrapper for HTTPS and Basic authentication depending
on the configuration.

Signed-off-by: Robin Jarry <rjarry@redhat.com>