pmd: add dpdk threads context switches metrics
The PMD threads should avoid context switching as much as possible. For
vhost user, they must do so from time to time since the virtio side
expects a kick on a file descriptor which requires to write() on it.

However, non-voluntary context switches should *NEVER* happen. When they
do it means there is an issue with isolation of the CPUs dedicated for
OVS DPDK.

Add two new metrics to expose both the voluntary and non-voluntary
context switches occurring on OVS DPDK PMD threads.

Signed-off-by: Robin Jarry <rjarry@redhat.com>