config: add new ovs-procdir setting
When running the exporter in a container, the PID namespace is different
and /proc only contains "1" (the process of the exporter). This leads to
an error when collecting context switches stats:

    ERROR readdir: open /proc/143795/task: no such file or directory

Add a setting to specify where the /proc folder of the PID namespace of
ovs-vswitchd (usually the base namespace) is mounted. By default, assume
everything is running in the same namespace.

Reported-by: Aaron Smith <aasmith@redhat.com>
Signed-off-by: Robin Jarry <rjarry@redhat.com>